Allow publisher to always publish by setting the threshold to zero.
On documentation (https://github.com/PickNikRobotics/topic_based_ros2_control/blob/main/doc/user.md)
trigger_joint_command_threshold can be configuration with zero so that the allow publisher to
always send the joint command.

However, in the code the condition was set with:
```
  if (diff <= trigger_joint_command_threshold_)
   {
     return hardware_interface::return_type::OK;
   }
```
In case trigger_joint_command_threshold_ is set to zero, diff can still be zero so
it will not trigger the send.

This PR tries to address the above mentioned case.